M24CSE032_PA2 (1)/PA2_submission/requirements.txt
torch>=2.0
torchaudio>=2.0
openai-whisper
speechbrain
librosa>=0.10,<0.11
soundfile
numpy
scipy
epitran
praat-parselmouth
dtw-python
TTS>=0.22
noisereduce
resemblyzer
jiwer
setuptools<71
